Replaced monodevelop dependency with other dot-net libraries
Without the external monodevelop, the gentoo system uninstalls many packages.
It is necedsary to have mono-addins to allow the monodevelop to start, but it complains about missing gnome-sharp
The editor doesn't actually work with just gnome-sharp, it requires gconf-sharp (which includes gnome-sharp itself)

There were two other files uninstalled by the removal of monodevelop, xsp and ctags. I have not yet determined whether these are relevant or not.